Update criteria for need_revision.
Proposals that need revision are those that do not have Accept as their
recommendation. The page also lists (and highlights) proposals with
missing or reviews that are not based on the latest submission.
Fix the comment at the top of review_status.php to give the correct name
for self.