Add effective date to events.
Added the ED column to the events table, and added a column for it
in the event editor; made the event editor wider to accomodate it.
Changed text in proposal display to reflect it.

This is because the effective date the registrar uses when fixing
the CUNYfirst data is important, but bears no relevance to the date
of the "event" -- when the registrar performed the fix.