Formalised the 'explain' query variable. But ...
Added code so that if form is submitted, the values will prepopulate
the form settings when it is displayed. This was intended as a way to
handle switching to/from the explain mode, but that is handled with a
regular link: the form never gets submitted. Next step is either to let
the form get submitted or to just leave it as it is and make the form
get reset if the user clicks on the ?explain link.