Add COPT4 to lists.
LPS, FCC, LIT, LANG, and SCI courses count for the fourth college
option group, in addition to SYN.

Updated curric.proposal_classes and curric.proposal_types to provide
the COPT4 class and type. The class is "not active" so people
will not be able to submit proposals for this dummy proposal_type.